--- Jeu/Src/ComponentsManagement/IAComponent/IAComponent.cpp
@@ -1,1 +1,0 @@
-#include "IAComponent.h"
