--- Jeu/Src/ComponentsManagement/IAComponent/IAs/SimpleFollow.h
@@ -1,16 +1,0 @@
-#pragma once
-
-#include "../IAs.h"
-
-class SimpleFollow : public IAs
-{
-public:
-	SimpleFollow() {}
-
-	void init(Entity* owner, Entity* target)
-	{
-		IAs::init(owner, target);
-	}
-
-	void update() override;
-};
