+++ Jeu/Src/ComponentsManagement/IASystem/IAs.cpp
@@ -1,6 +1,6 @@
 #include "IAs.h"
 
-Vector2D IAs::getDistanceToTarget()
+Vector2D IABase::getDistanceToTarget()
 {
 	Vector2D dist = Vector2D({ transform->position.x - target->getComponent<TransformComponent>().position.x, transform->position.y - target->getComponent<TransformComponent>().position.y });
 	return dist;
