+++ Jeu/Src/ComponentsManagement/IASystem/IAs/SimpleFollow.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include "../IAs.h"
+
+class SimpleFollow : public IABase
+{
+public:
+	SimpleFollow() {}
+
+	void update() override;
+};
